Move test_ethereum module for reuse in contract tests
